Implement q_remove_head & q_remove_tail
Implement this two functions to remove the head or tail
of the element and copy the value of the element to the 'sp'.

Using list_first_entry or list_last_entry to get the head or
tail element, then using 'list_del' to unlink the element from
the list.

At last, if 'sp' is non-NULL, copy the removed value to 'sp'.

Change-Id: I2a66c9a8d4d433f7e92ae2f988c67e72b5b89dc3