Implement q_delete_dup
Implement q_delete_dup to delete all nodes that have duplicate
string.
This function uses list_for_each_entry_safe to traverse the
queue, and uses a boolean variable to record current node
duplicate or not.
If the value of current node is the same as the next node,
this function will unlink and delete it, then assign true to
the boolean variable.
If the current node is not duplicate with the next and the
boolean variable is true, it will also delete the current node
then set the variable to false.

Change-Id: I2ef52480ba67cfa97e443a7d195af146ef15baa1