Implement swap and q_swap
Implement swap to swap the value pointers of two nodes.
Implement q_swap to swap every two adjacent nodes in queue.

The q_swap function uses list_for_each_entry_safe to traverse
queue.If the safe is not the head, it will skip the next node to
ensure that only every pair of node will be swapped.

Change-Id: Id141a8bd6b90c9e6184f08a76aa945bdb822eef3